+++ rbx_types/src/axes.rs
@@ -62,30 +62,6 @@ impl Axes {
     fn len(self) -> usize {
         self.bits().count_ones() as usize
     }
-
-    /// Whether the X axis, left face, and right face are enabled/included.
-    #[cfg(feature = "mlua")]
-    #[must_use]
-    #[doc(alias("X", "Left", "Right"))]
-    pub const fn x(self) -> bool {
-        self.contains(Self::X)
-    }
-
-    /// Whether the Y axis, top face, and bottom face are enabled/included.
-    #[cfg(feature = "mlua")]
-    #[must_use]
-    #[doc(alias("Y", "Top", "Bottom"))]
-    pub const fn y(self) -> bool {
-        self.contains(Self::Y)
-    }
-
-    /// Whether the Z axis, back face, and front face are enabled/included.
-    #[cfg(feature = "mlua")]
-    #[must_use]
-    #[doc(alias("Z", "Back", "Front"))]
-    pub const fn z(self) -> bool {
-        self.contains(Self::Z)
-    }
 }
 
 impl fmt::Debug for Axes {
@@ -246,101 +222,3 @@ mod serde_test {
         assert_eq!(all, de_all);
     }
 }
-
-#[cfg(feature = "mlua")]
-mod mlua_impl {
-    use super::*;
-
-    use std::fmt;
-
-    use mlua::{prelude::*, Variadic};
-
-    impl fmt::Display for Axes {
-        // This is not correct right now, because the `Lister` utility is not
-        // working as intended, but if PR #491 gets merged, this will then
-        // be accurate to the Roblox `tostring()`.
-        fn fmt(&self, out: &mut fmt::Formatter) -> fmt::Result {
-            let mut list = Lister::new();
-
-            if self.contains(Self::X) {
-                list.write(out, "X")?;
-            }
-
-            if self.contains(Self::Y) {
-                list.write(out, "Y")?;
-            }
-
-            if self.contains(Self::Z) {
-                list.write(out, "Z")?;
-            }
-
-            Ok(())
-        }
-    }
-
-    impl LuaUserData for Axes {
-        fn add_fields<F: LuaUserDataFields<Self>>(fields: &mut F) {
-            fields.add_field_method_get("X", |_, axes| Ok(axes.x()));
-            fields.add_field_method_get("Y", |_, axes| Ok(axes.y()));
-            fields.add_field_method_get("Z", |_, axes| Ok(axes.z()));
-            fields.add_field_method_get("Top", |_, axes| Ok(axes.y()));
-            fields.add_field_method_get("Bottom", |_, axes| Ok(axes.y()));
-            fields.add_field_method_get("Left", |_, axes| Ok(axes.x()));
-            fields.add_field_method_get("Right", |_, axes| Ok(axes.x()));
-            fields.add_field_method_get("Back", |_, axes| Ok(axes.z()));
-            fields.add_field_method_get("Front", |_, axes| Ok(axes.z()));
-        }
-
-        fn add_methods<M: LuaUserDataMethods<Self>>(methods: &mut M) {
-            methods.add_meta_method(LuaMetaMethod::Eq, |_, axes, other: LuaUserDataRef<Axes>| {
-                Ok(axes.flags == other.flags)
-            });
-            methods.add_meta_method(LuaMetaMethod::ToString, |_, axes, ()| Ok(axes.to_string()));
-        }
-    }
-
-    impl Axes {
-        fn into_lua(lua: &Lua) -> LuaResult<LuaTable> {
-            use crate::EnumItem;
-
-            let new = |_: &Lua, args: Variadic<EnumItem>| -> LuaResult<Axes> {
-                let mut flags: AxisFlags = AxisFlags::empty();
-
-                for arg in args {
-                    match &arg.ty[..] {
-                        "Axis" => {
-                            match arg.value {
-                                /* X */ 0_u32 => flags |= AxisFlags::X,
-                                /* Y */ 1_u32 => flags |= AxisFlags::Y,
-                                /* Z */ 2_u32 => flags |= AxisFlags::Z,
-                                _ => (),
-                            }
-                        }
-                        "NormalId" => {
-                            match arg.value {
-                                /* Right */ 0_u32 => flags |= AxisFlags::X,
-                                /* Top */ 1_u32 => flags |= AxisFlags::Y,
-                                /* Back */ 2_u32 => flags |= AxisFlags::Z,
-                                /* Left */ 3_u32 => flags |= AxisFlags::X,
-                                /* Bottom */ 4_u32 => flags |= AxisFlags::Y,
-                                /* Front */ 5_u32 => flags |= AxisFlags::Z,
-                                _ => (),
-                            }
-                        }
-                        _ => (),
-                    }
-                }
-
-                Ok(Axes { flags })
-            };
-
-            let mut axes = lua.create_table()?;
-            // axes.set("new", lua.create_function(new)?)?;
-            axes.set_readonly(true);
-            Ok(axes)
-        }
-    }
-}
-
-#[cfg(all(test, feature = "mlua"))]
-mod mlua_test {}
